% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quarters.R
\name{generate_quarters}
\alias{generate_quarters}
\title{Generate character vector of quarters}
\usage{
generate_quarters(year)
}
\arguments{
\item{year}{A numeric value for the year}
}
\value{
A character vector
}
\description{
Generate character vector of quarters
}
\examples{
generate_quarters(2026)
}
