% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_templates.R
\name{generate_templates}
\alias{generate_templates}
\title{Generate quarterly check-in templates}
\usage{
generate_templates(
  names,
  year = lubridate::year(Sys.Date()),
  output_dir = fs::path_wd()
)
}
\arguments{
\item{names}{A character vector of names}

\item{year}{\emph{Optional} A numeric value specifying the year. Defaults to current system year.}

\item{output_dir}{\emph{Optional} The folder location where the files will be saved. Defaults to current working directory.}
}
\value{
NULL. The function has not return value and it only used for its side-effects.
}
\description{
Generate quarterly check-in templates
}
\examples{
# No year specified. It will default to current year.
generate_templates(names = c("Peter", "Sam", "Lucy"))

# Year explicity specified.
generate_templates(names = c("Peter", "Sam", "Lucy"), year = 2026)
}
